+++ src/cspark/sdk/resources/_base.py
@@ -12,8 +12,7 @@ from .._config import Config
 from .._errors import SparkError
 from .._logger import get_logger
 from .._utils import get_retry_timeout, get_uuid, is_str_empty, sanitize_uri
-from .._version import about as sdk_info
-from .._version import sdk_logger, sdk_ua_header
+from .._version import about, sdk_ua_header
 
 __all__ = ['ApiResource', 'UriParams', 'Uri', 'HttpResponse']
 
@@ -22,7 +21,7 @@ class ApiResource:
     def __init__(self, config: Config):
         self.config = config
         self._client = Client()
-        self.logger = get_logger(sdk_logger, disable=not config.logger)
+        self.logger = get_logger(**config.logger.__dict__)
 
     def __enter__(self):
         return self
@@ -43,7 +42,7 @@ class ApiResource:
     def default_headers(self):
         return {
             **self.config.extra_headers,
-            'User-Agent': sdk_info,
+            'User-Agent': about,
             'x-spark-ua': sdk_ua_header,
             'x-request-id': get_uuid(),
             'x-tenant-name': self.config.base_url.tenant,